Explicitly close the Kafka consumer
Disposing the consumer is not sufficient for cleanly disconnecting the consumer.

From https://docs.confluent.io/clients-confluent-kafka-dotnet/current/overview.html#the-consume-loop:

Note that disposing the consumer instance after you are finished using it (achieved with the using block in the above example) will ensure that active sockets are closed and internal state is cleaned up. In order to leave the group cleanly - i.e. commit final offsets and trigger a group rebalance which ensures that any partitions owned by the consumer are re-assigned to other members in the group in a timely fashion - you additionally need to call the Close method prior to disposing.